+++ docs/make.jl
@@ -21,6 +21,7 @@ makedocs(
         "Dihderals and Ramachandran" => "dihedrals.md",
         "Hydrogen bonds" => "hydrogen_bonds.md",
         "Solvent Accessible Area" => "sasa.md",
+        "*m*-values" => "mvalue.md",
         "Auxiliary functions" => "auxiliary.md",
         "Examples" => "examples.md",
         "Tools" => "tools.md",