+++ src/tools/sasa.jl
@@ -193,7 +193,7 @@ are functions:
 
 By default, `atom_type = PDBTools.element`, a function that just returns the element symbol of the atom,
 and `atom_radius_from_type` obtains the vdW radius from the `PDBTools.elements` list given the element symbol.
-Here, the atomc radii of https://en.wikipedia.org/wiki/Atomic_radii_of_the_elements_(data_page) are used. 
+Here, the atomic radii of https://en.wikipedia.org/wiki/Atomic_radii_of_the_elements_(data_page) are used. 
 Atoms with missing radius have a `NaN` value, and the computation will not return meaningful
 values. 
 
@@ -259,7 +259,6 @@ function sasa_particles(
                     x = SVector(atoms[i].x, atoms[i].y, atoms[i].z)
                     v = SVector(dot_cache_i.x[idot], dot_cache_i.y[idot], dot_cache_i.z[idot])
                     push!(dots_exposed_i, x + v)
-
                 end
             end
             push!(dots, dots_exposed_i)
@@ -302,8 +301,8 @@ julia> sasa(at_sasa[1]) # single atom SASA
 
 """
 function sasa end
-sasa(p::SASA) = sum(i -> p[i], eachindex(p))
-sasa(p::SASA{N,<:AbstractVector{<:PDBTools.Atom}}, sel::Function) where {N} = sum(p[i] for i in eachindex(p) if sel(p.particles[i]))
+sasa(p::SASA) = sum(i -> p[i], eachindex(p); init=0.f0)
+sasa(p::SASA{N,<:AbstractVector{<:PDBTools.Atom}}, sel::Function) where {N} = sum(p[i] for i in eachindex(p) if sel(p.particles[i]); init=0.f0)
 sasa(p::SASA{N,<:AbstractVector{<:PDBTools.Atom}}, sel::String) where {N} = sasa(p, Select(sel))
 
 @testitem "sasa" begin