+++ CHANGELOG.md
@@ -10,8 +10,10 @@ PDBTools.jl Changelog
 [badge-fix]: https://img.shields.io/badge/Fix-purple.svg
 [badge-info]: https://img.shields.io/badge/Info-gray.svg
 
-Version 3.9.1-DEV
+Version 3.10.0-DEV
 -------------
+- ![FEATURE][badge-feature] add `mvalue` function for computing *m*-values from protein structures.
+- ![FEATURE][badge-feature] `in` is part of supported interface for iterators (e. g. `atoms[1] in residues[1]`)
 
 Version 3.9.0
 -------------